requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
beautifulsoup4
pandas
requests
faiss-cpu
sentence-transformers
openpyxl
lxml